update to world cup 2018

+++ src/main/java/org/jgayoso/ncomplo/business/services/UserService.java
@@ -6,6 +6,7 @@ import java.util.Locale;
 
 import org.apache.commons.lang3.RandomStringUtils;
 import org.jasypt.util.password.PasswordEncryptor;
+import org.jgayoso.ncomplo.business.entities.Invitation;
 import org.jgayoso.ncomplo.business.entities.League;
 import org.jgayoso.ncomplo.business.entities.User;
 import org.jgayoso.ncomplo.business.entities.User.UserComparator;
@@ -74,11 +75,11 @@ public class UserService {
 	public User registerFromInvitation(final Integer invitationId, final String login, final String name, final String email,
 			final Integer leagueId, final String password) {
 		
-		final boolean userExists = this.userRepository.exists(login);
+		final User existentUser = this.find(login);
         
-		if (userExists) {
-			//TODO change it
-			throw new InternalErrorException("User already exists");
+		if (existentUser != null) {
+			this.acceptInvitation(invitationId, leagueId, existentUser);
+			return existentUser;
 		}
 		final String hashedNewPassword = 
 				this.passwordEncryptor.encryptPassword(password);
@@ -96,8 +97,24 @@ public class UserService {
         newUser.getLeagues().add(league);
         league.getParticipants().add(newUser);
         
-        this.invitationRepository.delete(invitationId);
+        final Invitation invitation = this.invitationRepository.findOne(invitationId);
+        if (invitation.getToken() == null) { 
+        	this.invitationRepository.delete(invitationId);
+        }
         return newUser;
+	}
+	
+	@Transactional
+    public void acceptInvitation(final Integer invitationId, final Integer leagueId, final User user) {
+	    
+	    final League league = this.leagueRepository.findOne(leagueId);
+	    user.getLeagues().add(league);
+        league.getParticipants().add(user);
+        
+        final Invitation invitation = this.invitationRepository.findOne(invitationId);
+        if (invitation.getToken() == null) { 
+        	this.invitationRepository.delete(invitationId);
+        }
 	}
     
     @Transactional
